# api-contract-v1.yaml
openapi: 3.1.0
info:
  title: Event Face Matching API
  version: 1.0.0
  description: |
    Phase 1 contract — single studio, no billing. Multi-tenant structures are
    present but only one studio row exists.

    ## Who owns what
    Backend (routes, DB, worker) is implemented by the studio owner.
    Frontend (three React apps) is implemented against this spec via a mock server.
    Neither side blocks the other. When the real server is live, only the base URL changes.

    ## Two separate auth contexts — do not mix them
    - **Staff** (studio owners, photographers, platform admins) use
      `Authorization: Bearer <access_token>`.
    - **Guests** have no account. They use `X-Guest-Session: <session_token>`,
      scoped to exactly one event, and can reach only the `/g/*` routes.

    Keeping these in different headers is deliberate: it makes it structurally
    hard to accidentally accept a guest token on a staff route.

    ## photos.status state machine
    ```
    pending ──► uploaded ──► processing ──► done
                                 │
                                 └────────► failed ──► (retry) ──► processing
    ```
    - `pending`    row created, presigned URL issued, bytes not yet confirmed
    - `uploaded`   client confirmed the PUT succeeded; job enqueued
    - `processing` a worker has claimed the job
    - `done`       faces extracted, embedded, stored, matched against guests
    - `failed`     gave up after 5 attempts; `error_code` explains why

...

...
servers:
...
security:
  - staffAuth: []
paths:
...
  /events/{event_id}/uploads:
    parameters:
      - { $ref: '#/components/parameters/EventId' }
    post:
      tags: [uploads]
      summary: Request presigned upload URLs for a batch of photos
      description: |
        The client describes the files it intends to upload; the server
        authorizes, generates the object keys itself, and returns one
        presigned PUT URL per file.

        **The client never chooses the object key.** If it did, a caller could
        overwrite another event's photos by supplying a crafted key.

        The signature constrains bucket, key, method, `content_type` and
        `max_bytes`, and expires at `expires_at` (~10 minutes).

        **Presign in rolling chunks, not all at once.** A photographer may drop
        600 files in a single round. Those live in a client-side queue with
        4-6 uploads in flight, which can take 20+ minutes on venue wifi. If all
        600 URLs were minted upfront, the tail would expire before the queue
        reached it. The uploader therefore calls this endpoint repeatedly for
        the next slice of its queue, keeping every URL under a minute old when
        it is used.

        Server cap is 250 files per call. The uploader hides this entirely;
        photographers drop as many files as they like.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  minItems: 1
                  maxItems: 250
                  items:
                    type: object
                    required: [client_ref, content_type, size_bytes]
                    properties:
                      client_ref:
                        type: string
                        description: |
                          Client-generated id echoed back in the response so the
                          uploader can match URLs to its own queue entries.
                          Never used as, or incorporated into, the object key.
                        maxLength: 100
                      content_type: { type: string, enum: [image/jpeg, image/png, image/webp] }
                      size_bytes: { type: integer, minimum: 1, maximum: 26214400 }
      responses:
        '201':
          description: URLs issued
          content:
            application/json:
              schema:
                type: object
                required: [uploads]
                properties:
                  uploads:
                    type: array
                    items:
                      type: object
                      required: [client_ref, photo_id, upload_url, expires_at]
                      properties:
                        client_ref: { type: string }
                        photo_id: { type: string, format: uuid }
                        upload_url: { type: string, format: uri }
                        expires_at: { type: string, format: date-time }
        '403':
          description: Event is not active, so uploads are refused
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example:
                error:
                  code: EVENT_NOT_ACTIVE
                  message: This event has not been activated.
        '422': { $ref: '#/components/responses/ValidationError' }
  /events/{event_id}/photos/complete:
    parameters:
      - { $ref: '#/components/parameters/EventId' }
    post:
      tags: [uploads]
      summary: Confirm which uploads succeeded, and enqueue them
      description: |
        Called after the PUTs finish. Report only the photos whose upload
        actually succeeded — failed ones stay `pending` and are swept up later.

        Moves each photo `pending → uploaded` and inserts a processing job in
        the **same transaction**, so a photo can never be marked uploaded
        without a job to process it.

        Idempotent: re-sending a `photo_id` already `uploaded` or beyond is a
        no-op, not an error. The uploader retries on flaky networks and must be
        able to do so safely.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [photo_ids]
              properties:
                photo_ids:
                  type: array
                  minItems: 1
                  maxItems: 250
                  items: { type: string, format: uuid }
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [enqueued, skipped]
                properties:
                  enqueued: { type: integer }
                  skipped:
                    type: integer
                    description: Already past `pending`; ignored.
...
  /photos/{photo_id}/retry:
    parameters:
      - name: photo_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [photos]
      summary: Re-enqueue a failed photo
      description: Resets attempts to 0 and returns the photo to `uploaded`.
      responses:
        '200':
          description: Re-enqueued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
        '409':
          description: Photo is not in `failed` state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ───────────────────────────── guest ─────────────────────────────
...
components:
  securitySchemes:
    staffAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Studio owners, photographers and platform admins.
    guestAuth:
      type: apiKey
      in: header
      name: X-Guest-Session
      description: Opaque session token, scoped to one event, valid for `/g/*` only.
  parameters:
    EventId:
      name: event_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable identifier. Frontend switches on this, never on `message`.
            message:
              type: string
              description: Human-readable text, safe to show a user.
            details:
              type: object
              additionalProperties: true
...
    Photo:
      type: object
      required: [id, status, created_at]
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [pending, uploaded, processing, done, failed] }
        thumbnail_url:
          type: [string, 'null']
          format: uri
          description: Presigned GET, short-lived. Null until processed.
        face_count: { type: [integer, 'null'] }
        error_code:
          type: [string, 'null']
          description: Present only when status is `failed`.
        created_at: { type: string, format: date-time }
        processed_at: { type: [string, 'null'], format: date-time }
...